+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Added `FullscreenSystem` to handle fullscreen events and pointer lock.
+- `PointerSystem` supports pointer lock through the `FullscreenSystem`.
 ### Changed
 - `KeyboardSystem` messages emitted use JS `KeyboardEvent.code` (physical keyboard) rather than `KeyboardEvent.key` (input character).
 - `System` to store `SystemEntities` as a map rather than as an array, for easier random access.